AC: use a RootDirectoryDigest for directory assets
Which means we avoid needing to calculate TreeDigest`s whilst still
creating Actions/ActionResults that encode that an asset is a directory.

Note: the bb-remote-asset server does not validate that the
RootDirectoryDigest actually references a Directory proto because
bb-remote-asset treats the digest as an opaque digest.

See https://github.com/buildbarn/bb-remote-asset/pull/34#issuecomment-2220546603